requirements.txt
rasa==3.0.0
rasa-sdk==3.0.0
lingua-language-detector==1.1.1
requests==2.31.0
geocoder==1.38.1
